+++ adaptor/typed/typedjournal/bounds_test.go
@@ -1,21 +1,24 @@
-package journal_test
+package typedjournal_test
 
 import (
-	"bytes"
 	"context"
 	"testing"
 	"time"
 
+	. "github.com/dogmatiq/persistencekit/adaptor/typed/typedjournal"
+	"github.com/dogmatiq/persistencekit/adaptor/typed/typedmarshaler"
 	"github.com/dogmatiq/persistencekit/driver/memory/memoryjournal"
-	. "github.com/dogmatiq/persistencekit/journal"
 )
 
 func TestIsFresh(t *testing.T) {
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	defer cancel()
 
-	store := &memoryjournal.Store{}
-	j, err := store.Open(ctx, "test")
+	store := Store[int, typedmarshaler.JSON[int]]{
+		Store: &memoryjournal.Store{},
+	}
+
+	j, err := store.Open(ctx, "<name>")
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -34,7 +37,7 @@ func TestIsFresh(t *testing.T) {
 	})
 
 	t.Run("when the journal has records", func(t *testing.T) {
-		if err := j.Append(ctx, 0, []byte("<record-0>")); err != nil {
+		if err := j.Append(ctx, 0, 100); err != nil {
 			t.Fatal(err)
 		}
 
@@ -70,8 +73,11 @@ func TestIsEmpty(t *testing.T) {
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	defer cancel()
 
-	store := &memoryjournal.Store{}
-	j, err := store.Open(ctx, "test")
+	store := Store[int, typedmarshaler.JSON[int]]{
+		Store: &memoryjournal.Store{},
+	}
+
+	j, err := store.Open(ctx, "<name>")
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -90,7 +96,7 @@ func TestIsEmpty(t *testing.T) {
 	})
 
 	t.Run("when the journal has records", func(t *testing.T) {
-		if err := j.Append(ctx, 0, []byte("<record-0>")); err != nil {
+		if err := j.Append(ctx, 0, 100); err != nil {
 			t.Fatal(err)
 		}
 
@@ -126,8 +132,11 @@ func TestFirstRecord(t *testing.T) {
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	defer cancel()
 
-	store := &memoryjournal.Store{}
-	j, err := store.Open(ctx, "test")
+	store := Store[int, typedmarshaler.JSON[int]]{
+		Store: &memoryjournal.Store{},
+	}
+
+	j, err := store.Open(ctx, "<name>")
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -146,11 +155,11 @@ func TestFirstRecord(t *testing.T) {
 	})
 
 	t.Run("when the journal has records", func(t *testing.T) {
-		if err := j.Append(ctx, 0, []byte("<record-0>")); err != nil {
+		if err := j.Append(ctx, 0, 100); err != nil {
 			t.Fatal(err)
 		}
 
-		if err := j.Append(ctx, 1, []byte("<record-1>")); err != nil {
+		if err := j.Append(ctx, 1, 200); err != nil {
 			t.Fatal(err)
 		}
 
@@ -166,8 +175,8 @@ func TestFirstRecord(t *testing.T) {
 				t.Fatalf("unexpected position: got %d, want %d", pos, 0)
 			}
 
-			expect := []byte("<record-0>")
-			if !bytes.Equal(rec, expect) {
+			expect := 100
+			if rec != expect {
 				t.Fatalf("unexpected record: got %q, want %q", rec, expect)
 			}
 		})
@@ -190,8 +199,8 @@ func TestFirstRecord(t *testing.T) {
 				t.Fatalf("unexpected position: got %d, want %d", pos, 1)
 			}
 
-			expect := []byte("<record-1>")
-			if !bytes.Equal(rec, expect) {
+			expect := 200
+			if rec != expect {
 				t.Fatalf("unexpected record: got %q, want %q", rec, expect)
 			}
 		})
@@ -218,8 +227,11 @@ func TestLastRecord(t *testing.T) {
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	defer cancel()
 
-	store := &memoryjournal.Store{}
-	j, err := store.Open(ctx, "test")
+	store := Store[int, typedmarshaler.JSON[int]]{
+		Store: &memoryjournal.Store{},
+	}
+
+	j, err := store.Open(ctx, "<name>")
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -238,11 +250,11 @@ func TestLastRecord(t *testing.T) {
 	})
 
 	t.Run("when the journal has records", func(t *testing.T) {
-		if err := j.Append(ctx, 0, []byte("<record-0>")); err != nil {
+		if err := j.Append(ctx, 0, 100); err != nil {
 			t.Fatal(err)
 		}
 
-		if err := j.Append(ctx, 1, []byte("<record-1>")); err != nil {
+		if err := j.Append(ctx, 1, 200); err != nil {
 			t.Fatal(err)
 		}
 
@@ -258,8 +270,8 @@ func TestLastRecord(t *testing.T) {
 				t.Fatalf("unexpected position: got %d, want %d", pos, 1)
 			}
 
-			expect := []byte("<record-1>")
-			if !bytes.Equal(rec, expect) {
+			expect := 200
+			if rec != expect {
 				t.Fatalf("unexpected record: got %q, want %q", rec, expect)
 			}
 		})
@@ -282,8 +294,8 @@ func TestLastRecord(t *testing.T) {
 				t.Fatalf("unexpected position: got %d, want %d", pos, 1)
 			}
 
-			expect := []byte("<record-1>")
-			if !bytes.Equal(rec, expect) {
+			expect := 200
+			if rec != expect {
 				t.Fatalf("unexpected record: got %q, want %q", rec, expect)
 			}
 		})